Moves some dupe BBCode to core.
Sub, sup, icon, eicon, and url now only exist in core. User was left in place as the profile/chat versions display differently but I can pull that in if it is wanted.
